Move gson model to gson/dvach.
