+++ data/src/main/kotlin/dvachmovie/api/model/gson/dvach/catalog/FilesItem.kt
@@ -1,4 +1,4 @@
-package dvachmovie.api.model.catalog
+package dvachmovie.api.model.gson.dvach.catalog
 
 import com.google.gson.annotations.SerializedName
